requirements.txt
uvicorn==0.30.3
fastapi==0.111.0
langchain==0.2.11
langchain_community==0.2.10
langchain_openai==0.1.19
langchain_huggingface==0.0.3
sentence_transformers==3.0.1
pypdf==4.3.1
pytest==8.3.2
faiss_cpu==1.8.0
python-dotenv~=1.0.1
pydantic~=2.8.2
sqlalchemy